Fix circular imports.
