+++ gsadjust/gui/__init__.py
@@ -19,4 +19,4 @@ constitute any such warranty. The software is provided on the condition that
 neither the USGS nor the U.S. Government shall be held liable for any damages
 resulting from the authorized or unauthorized use of the software.
 """
-from . import dialogs, menus, messages, table, tabs, widgets
+from . import dialogs, menus, messages, tabs, utils, widgets
