+++ gsadjust/gui/messages.py
@@ -44,57 +44,3 @@ def show_message(message, title, icon=QtWidgets.QMessageBox.Warning, helptext=No
     msg.addButton(QtWidgets.QPushButton("OK"), QtWidgets.QMessageBox.RejectRole)
     msg.show()
     return msg
-
-
-class DialogMeterType(QtWidgets.QMessageBox):
-    def __init__(self):
-        super(DialogMeterType, self).__init__()
-        self.setText("Choose meter file to import")
-        self.addButton(
-            QtWidgets.QPushButton(' CG-3, CG-5 '), QtWidgets.QMessageBox.YesRole
-        )
-        self.addButton(QtWidgets.QPushButton(' CG-6 '), QtWidgets.QMessageBox.YesRole)
-        self.addButton(
-            QtWidgets.QPushButton(' CG-6 Tsoft '), QtWidgets.QMessageBox.YesRole
-        )
-        self.addButton(QtWidgets.QPushButton(' Burris '), QtWidgets.QMessageBox.YesRole)
-        self.addButton(QtWidgets.QPushButton(' CSV '), QtWidgets.QMessageBox.YesRole)
-        self.addButton(
-            QtWidgets.QPushButton(' Cancel '), QtWidgets.QMessageBox.RejectRole
-        )
-        self.buttonClicked.connect(self.onClicked)
-        self.setWindowModality(QtCore.Qt.ApplicationModal)
-
-    def onClicked(self, btn):
-        meter = {
-            ' CG-3, CG-5 ': 'CG5',
-            ' CG-6 ': 'CG6',
-            ' CG-6 Tsoft ': 'CG6Tsoft',
-            ' Burris ': 'Burris',
-            ' CSV ': 'csv',
-            ' Cancel ': 'Cancel',
-        }
-        self.meter_type = meter[btn.text()]
-        if self.meter_type == 'Cancel':
-            self.reject()
-        else:
-            self.accept()
-
-
-class DialogOverwrite(QtWidgets.QMessageBox):
-    def __init__(self):
-        super(DialogOverwrite, self).__init__()
-        self.overwrite = False
-        self.setText("Overwrite existing data?")
-        self.setWindowTitle("GSadjust")
-        ow_button = QtWidgets.QPushButton('Overwrite')
-        ow_button.clicked.connect(self.onClicked)
-
-        cancel_button = QtWidgets.QPushButton('Cancel')
-        cancel_button.clicked.connect(self.close)
-        self.addButton(cancel_button, 0)
-        self.addButton(ow_button, 1)
-        self.setWindowModality(QtCore.Qt.ApplicationModal)
-
-    def onClicked(self, btn):
-        self.accept()
